refactor(challenge): ChallengeDifferent/PortfolioOneOnOne 슬림화
하드코딩 데이터·프레젠테이션 블록을 co-located 파일로 추출:
- ChallengeDifferent 350→136 (challenge-different/: BenefitSection, DifferentSection, constants)
- PortfolioOneOnOne 334→207 (portfolio-one-on-one/: Card, data)
순수 이동, 동작 변경 없음. tsc 0 / lint 통과.
