+++ apps/web/src/domain/program/challenge/challenge-view/ChallengeDifferent.tsx
@@ -1,14 +1,9 @@
-import benefitImg1 from '@/assets/benefit1.jpg';
-import benefitImg2 from '@/assets/benefit2.jpg';
-import PaybackTicket from '@/domain/program/challenge/challenge-view/PaybackTicket';
 import { challengeColors } from '@/domain/program/challenge/challengeColors';
-import BenefitCard from '@/domain/program/program-detail/different/BenefitCard';
-import DifferentCard, {
-  DifferentCardProps,
-} from '@/domain/program/program-detail/different/DifferentCard';
-import SuperTitle from '@/domain/program/program-detail/SuperTitle';
 import { ChallengeType, challengeTypeSchema } from '@/schema';
 import { useMemo } from 'react';
+import BenefitSection from './challenge-different/BenefitSection';
+import { getDifferentList } from './challenge-different/challenge-different.constants';
+import DifferentSection from './challenge-different/DifferentSection';
 
 const {
   PORTFOLIO,
@@ -18,31 +13,6 @@ const {
   PERSONAL_STATEMENT,
   ETC,
 } = challengeTypeSchema.enum;
-
-export const tripleBenefits = [
-  {
-    title: '챌린지 수료시, 챌린지 3종 할인 쿠폰 제공',
-    options: [
-      '미션 80점 이상 완료 시 자기소개서, 포트폴리오 완성, 면접 챌린지 등 할인 쿠폰 발급!\n(챌린지 3종 중 1회 적용 가능)',
-    ],
-    imgUrl: { src: '/images/benefit0.svg' },
-  },
-  {
-    title: '온라인 대시보드',
-    options: [
-      `학습부터 미션 수행까지 올인원으로 관리할 수 있는 전용 대시보드`,
-      `미션 현황과 기수별 주요 공지도 함께 열람할 수 있어 몰입감 UP!`,
-    ],
-    imgUrl: benefitImg1,
-  },
-  {
-    title: '프로그램 수료증 발급',
-    options: [
-      `프로그램 종료 시, 참여자분들께 렛츠커리어에서 인증하는 참여 수료증을 발급해드립니다.`,
-    ],
-    imgUrl: benefitImg2,
-  },
-];
 
 const ChallengeDifferent = ({
   challengeType,
@@ -97,96 +67,7 @@ const ChallengeDifferent = ({
     }
   }, [challengeType]);
 
-  const differentList: DifferentCardProps[] = isResumeTemplate
-    ? [
-        {
-          order: 1,
-          title: `취업 준비가 더 이상 막막하지 않도록\nA부터 Z까지 알려주는 학습 콘텐츠`,
-          options: [
-            '초보 취준생도 따라갈 수 있는 친절한 길라잡이',
-            '합격자 예시를 포함하여 서류 작성 스킬 UP',
-            '2025년 주요 기업/직무 합격 자료로 파악하는 채용 트렌드',
-          ],
-          imageUrl: {
-            desktop: '/challenge-detail/different/desktop/contents_desktop.gif',
-            mobile: '/challenge-detail/different/mobile/contents_mobile.gif',
-          },
-          styles,
-        },
-        {
-          order: 2,
-          title: `서류와 면접의 기초 베이스가 되어줄\n미션 템플릿으로 나만의 이력서 완성`,
-          options: [
-            '하루 30분, 경험 정리부터 이력서까지 완성하는 실습',
-            '누구나 쉽게 채울 수 있는 노션 템플릿',
-            '수료 후에도 자산으로 남는 경험 회고록',
-          ],
-          imageUrl: {
-            desktop: '/challenge-detail/different/desktop/template_desktop.gif',
-            mobile: '/challenge-detail/different/mobile/template_mobile.gif',
-          },
-          styles,
-        },
-        {
-          order: 3,
-          title: `함께라서 할 수 있어요.\n챌린지 참여자들과 함께 동기부여!`,
-          options: [
-            '취업 여정의 동료들과 함께 지치지 않고 완성해요.',
-            '자유로운 커뮤니티를 통해 스터디 함께할 수 있어요.',
-            '서류 작성을 통해 커리어 고민 함께 나눠요.',
-          ],
-          imageUrl: {
-            desktop:
-              '/challenge-detail/different/desktop/community_desktop.gif',
-            mobile: '/challenge-detail/different/mobile/community_mobile.gif',
-          },
-          styles,
-        },
-      ]
-    : [
-        {
-          order: 1,
-          title: `취업 준비가 더 이상 막막하지 않도록\nA부터 Z까지 알려주는 학습 콘텐츠`,
-          options: [
-            '초보 취준생도 따라갈 수 있는 친절한 길라잡이',
-            '합격자 예시를 포함하여 콘텐츠 이해도 UP',
-            'PDF 30페이지 분량의 추가 콘텐츠 제공',
-          ],
-          imageUrl: {
-            desktop: '/challenge-detail/different/desktop/contents_desktop.gif',
-            mobile: '/challenge-detail/different/mobile/contents_mobile.gif',
-          },
-          styles,
-        },
-        {
-          order: 2,
-          title: `서류와 면접의 기초 베이스가 되어줄\n미션 템플릿으로 나만의 취업 가이드북 완성`,
-          options: [
-            '하루 30분, 서류를 완성하는 실습',
-            '누구나 쉽게 채울 수 있는 노션 템플릿',
-            '수료 후에도 자산으로 남는 취업 가이드북',
-          ],
-          imageUrl: {
-            desktop: '/challenge-detail/different/desktop/template_desktop.gif',
-            mobile: '/challenge-detail/different/mobile/template_mobile.gif',
-          },
-          styles,
-        },
-        {
-          order: 3,
-          title: `시간은 어느새 흐르고,\n혼자 하기 어렵다면 사람들과\n함께 공유하며 성장하는 동기부여 시스템`,
-          options: [
-            `사람들과 함께 공유하며 성장하는\n오픈 카톡 커뮤니티`,
-            '미션 80점 이상 완료 시, 수료증 지급',
-          ],
-          imageUrl: {
-            desktop:
-              '/challenge-detail/different/desktop/community_desktop.gif',
-            mobile: '/challenge-detail/different/mobile/community_mobile.gif',
-          },
-          styles,
-        },
-      ];
+  const differentList = getDifferentList(isResumeTemplate, styles);
 
   const paypackImgSrc = (() => {
     if (isResumeTemplate) return '/images/payback-career-start157.png';
@@ -230,119 +111,24 @@ const ChallengeDifferent = ({
       className="flex w-full flex-col gap-y-[70px] py-16 md:gap-y-40 md:py-40"
     >
       {/* 차별점 */}
-      <div className="flex w-full flex-col gap-y-8 md:gap-y-20">
-        <div className="flex w-full flex-col gap-y-6 md:gap-y-12">
-          <SuperTitle style={{ color: styles.primaryColor }}>차별점</SuperTitle>
-          <div className="flex flex-col gap-y-3 md:items-center">
-            <p
-              className="text-xsmall16 md:text-small18 font-bold"
-              style={{ color: styles.primaryColor }}
-            >
-              {isResumeTemplate
-                ? `${challengeTitle}에서 얻어갈 수 있는 것들`
-                : '비교 불가!'}
-            </p>
-            <div className="md:text-xlarge28 whitespace-pre text-[22px] font-bold text-black md:text-center">
-              {isResumeTemplate ? (
-                <>
-                  <span>렛츠커리어 챌린지만의 차별점</span>
-                  <br />
-                </>
-              ) : (
-                <>
-                  <span>{challengeTitle}만의</span>
-                  <br />
-                </>
-              )}
-              <span>
-                {isResumeTemplate ? '이 모든걸 ' : '차별점, 이 모든걸 '}
-                <img
-                  className="inline-block h-auto w-8 md:w-10"
-                  src={`/icons/${iconName}`}
-                  alt=""
-                />{' '}
-                얻어가실 수 있어요!
-              </span>
-            </div>
-          </div>
-        </div>
-        <div className="flex w-full flex-col gap-y-6">
-          {differentList.map((different) => (
-            <DifferentCard
-              key={different.title}
-              order={different.order}
-              title={different.title}
-              options={different.options}
-              imageUrl={different.imageUrl}
-              styles={styles}
-            />
-          ))}
-          {deposit >= 10000 && (
-            <div
-              className="text-small18 md:text-medium22 relative flex w-full gap-x-2 overflow-hidden rounded-md px-5 pb-10 pt-[30px] font-bold md:px-10 md:py-[50px]"
-              style={{
-                backgroundColor: styles.primaryLightColor,
-                color: styles.primaryLightColor,
-              }}
-            >
-              <span style={{ color: styles.primaryColor }}>혜택</span>
-              <p className="z-10 whitespace-pre text-black">
-                모든 커리큘럼을 따라오기만 하면,
-                <br className="md:hidden" /> {deposit / 10000}
-                만원을 페이백해드려요!
-              </p>
-              {challengeType === PERSONAL_STATEMENT ? (
-                <PaybackTicket
-                  deposit={deposit}
-                  className="absolute bottom-0 right-0 h-auto w-28 md:top-0 md:w-48"
-                />
-              ) : (
-                <img
-                  className="absolute bottom-0 right-0 h-auto w-28 md:top-0 md:w-48"
-                  src={paypackImgSrc}
-                  alt={`페이백 ${deposit / 10000}만원`}
-                />
-              )}
-            </div>
-          )}
-        </div>
-      </div>
+      <DifferentSection
+        challengeType={challengeType}
+        challengeTitle={challengeTitle}
+        deposit={deposit}
+        isResumeTemplate={isResumeTemplate}
+        iconName={iconName}
+        paypackImgSrc={paypackImgSrc}
+        differentList={differentList}
+        styles={styles}
+      />
 
       {/* 혜택 */}
-      <div className="flex w-full flex-col gap-y-8 md:items-center md:gap-y-16">
-        <p className="text-small20 md:text-xlarge28 whitespace-pre-line font-bold md:text-center">
-          여기서 끝이 아니죠
-          <br />
-          {challengeTitle}
-          <br className="md:hidden" /> 참여자만을 위한{' '}
-          {isResumeTemplate ? '' : '트리플'} 혜택!
-        </p>
-
-        <div
-          className="-mx-5 flex w-fit flex-col gap-y-4 overflow-x-auto px-5 md:-mx-10 md:px-10 lg:px-0"
-          style={{ color: styles.primaryColor }}
-        >
-          {(challengeType === EXPERIENCE_SUMMARY || challengeType === ETC) && (
-            <BenefitCard
-              title={tripleBenefits[0].title}
-              options={tripleBenefits[0].options}
-              imgUrl={tripleBenefits[0].imgUrl.src}
-            />
-          )}
-          <BenefitCard
-            title={tripleBenefits[1].title}
-            options={tripleBenefits[1].options}
-            imgUrl={tripleBenefits[1].imgUrl.src}
-          />
-          {!isResumeTemplate && (
-            <BenefitCard
-              title={tripleBenefits[2].title}
-              options={tripleBenefits[2].options}
-              imgUrl={tripleBenefits[2].imgUrl.src}
-            />
-          )}
-        </div>
-      </div>
+      <BenefitSection
+        challengeType={challengeType}
+        challengeTitle={challengeTitle}
+        isResumeTemplate={isResumeTemplate}
+        primaryColor={styles.primaryColor}
+      />
     </section>
   );
 };
